chore: update spec-update.only-add-missing-method-test-cases.spec.ts
spec-update.only-add-missing-method-test-cases.spec.ts
