+++ tests/spec/spec-update.only-add-missing-method-test-cases.spec.ts
@@ -1,59 +1,43 @@
+import { LogEntry } from '@angular-devkit/core/src/logger';
 import { Tree } from '@angular-devkit/schematics';
-import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 import { filter } from 'rxjs/operators';
-import { collectionPath } from './common';
+import { setupBase } from './common';
+
+const folder = 'only-add-missing-method-test-cases';
+const file = 'c.ts';
 
 describe('Calling update', () => {
     let tree = Tree.empty();
 
     beforeEach(() => {
         tree = Tree.empty();
-        tree.create(
-            'c.ts',
-            `import { LogService, BDep } from '@angular/core';
-
-            export class C  {
-                public methodOne() {}
-                methodTwo() {}
-                methodThree() {}
-            `
-        );
-
-        tree.create(
-            'c.spec.ts',
-            `import { bDep } from '@angular/core';
-
-describe('C', () => {
-    it('existing method one test', () => {
-        c.methodOne();
-    })
-    it('existing test case with name of methodTwo in the spec title', () => {
-        c.methodTwo();
-    })
-    it('just any name', () => {
-        c.methodThree();
-    })
-});`
-        );
     });
 
-    it('should work and not throw errors', async  () => {
+    it('should work and not throw errors', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
+        const { run, fullFileName, add, testFileName, letLogger } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName)
         // act
-        const errors = [];
-        runner.logger.pipe(filter(v => v.level === 'error')).subscribe(v => errors.push(v));
-        await runner.runSchematicAsync('spec', { name: './c.spec.ts', update: true }, tree).toPromise();
+        await run({ name: fullFileName, update: false });
+        // act
+        const errors: LogEntry[] = [];
+        letLogger(logger => logger.pipe(filter((v) => v.level === 'error')).subscribe((v) => errors.push(v)));
+        await run({ name: './c.spec.ts', update: true }).catch(e => {
+            fail('Failing the test on purpose: There should not be any errors thrown here. ')
+        });
         // assert
         expect(errors.length).toBe(0);
     });
 
-    it('should update the spec file', async  () => {
+    it('should update the spec file', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
+        const { run, fullFileName, add, testFileName } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName)
         // act
-        const result = await runner.runSchematicAsync('spec', { name: './c.spec.ts', update: true }, tree).toPromise();
-        const contents = result.readContent('./c.spec.ts');
+        const result = await run({ name: fullFileName, update: false });
+        const contents = result.readContent(testFileName);
         // assert
         expect(contents).toEqual(`import { bDep } from '@angular/core';
