+++ tests/spec/test-data/only-add-missing-method-test-cases/c.ts
@@ -1,0 +1,7 @@
+import { LogService, BDep } from '@angular/core';
+
+export class C  {
+    public methodOne() {}
+    methodTwo() {}
+    methodThree() {}
+}
